fix(home) automated recent beers re-rendering
